requirements.txt
APScheduler==3.6.3
attrs==21.4.0
cachetools==4.2.2
certifi==2022.5.18.1
emoji==1.7.0
ephem==4.1.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.0.1
pytest-sugar==0.9.4
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
termcolor==1.1.0
tomli==2.0.1
tornado==6.1
tzdata==2022.1
tzlocal==4.2
